+++ handler-annotations/src/main/java/com/ensarsarajcic/neovim/java/handler/annotations/NeovimNotificationHandler.java
@@ -44,8 +44,9 @@ import java.lang.annotation.Target;
 public @interface NeovimNotificationHandler {
     /**
      * Name of notification this method will handle
+     * By default fully qualified method name is used
      *
      * @return name of notification
      */
-    String value();
+    String value() default "";
 }