+++ handler-annotations/src/main/java/com/ensarsarajcic/neovim/java/handler/annotations/NeovimRequestListener.java
@@ -46,8 +46,9 @@ import java.lang.annotation.Target;
 public @interface NeovimRequestListener {
     /**
      * Name of request this method will handle
+     * By default fully qualified method name is used
      *
      * @return name of request
      */
-    String value();
+    String value() default "";
 }