+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+- autocommand events for `plugin-host` - `NeovimJavaPrepare` and `NeovimJavaReady`
+- fully qualified names as defaults for command/autocommand and request/notification handlers
+
 ## [0.4.1] - 2022-06-04
 
 - `plugin-host` module for both standalone and hosted plugins